+++ src/frontend/admin/src/services/api/models/ContractDefinition.ts
@@ -3,11 +3,23 @@ import { Optional } from "@/types/utils";
 export type ContractDefinition = {
   id: string;
   title: string;
-  description?: string;
-  body?: string;
-  appendix?: string;
+  description: string;
+  body: string;
+  appendix: string;
   language: string;
-  name: string;
+  name: ContractDefinitionTemplate;
+};
+
+export type ContractDefinitionFormValues = Omit<
+  ContractDefinition,
+  "id" | "name"
+> & {
+  name: ContractDefinitionTemplate | "";
 };
 
 export type DTOContractDefinition = Optional<ContractDefinition, "id">;
+
+export enum ContractDefinitionTemplate {
+  DEFAULT = "contract_definition_default",
+  UNICAMP = "contract_definition_unicamp",
+}